requirements.txt
discord.py>=1.2.3
python-dotenv==0.10.3
sqlalchemy